pyproject.toml
[tool.poetry]
name = "zazzle-agent"
version = "0.1.0"
description = "Zazzle Affiliate Agent - Automated product generation and Reddit interaction system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
alembic = "^1.13.0"
fastapi-users = "^12.0.0"
fastapi-mail = "^1.4.1"
fastapi-cache2 = "^0.2.1"
slowapi = "^0.1.9"
fastapi-pagination = "^0.12.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
pandas = "^2.2.1"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"
praw = "^7.7.0"
httpx = "^0.24.1"
pytest-mock = "^3.12.0"
Pillow = "^10.0.0"
schedule = "^1.2.0"
pytest-cov = "^6.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
] 